Configuracao planos

+++ backend/proofchain/src/main/java/com/proofchain/Dtos/request/NewInstituitionRequestDto.java
@@ -24,5 +24,5 @@ public class NewInstituitionRequestDto {
     private String email;
     @Size(min=8)
     private String password;
-    private int idPlan;
+    private Long idPlan;
 }
